Move backdrop back to pageblocks
