+++ src/pages/tv/[tvID]/season/[seasonID]/episode/[...episodeID].tsx
@@ -14,7 +14,7 @@ import EpisodeWatchButton from "../../../../../../components/watchButton/Episode
 import { AiFillStar } from "react-icons/ai";
 import WatchTrailerButton from "../../../../../../components/common/buttons/WatchTrailerButton";
 import JustWatch from "../../../../../../components/common/JustWatch";
-import Backdrop from "../../../../../../components/contentHeader/Backdrop";
+import Backdrop from "../../../../../../components/pageBlocks/Backdrop";
 
 const EpisodePage = () => {
   const router = useRouter();
